backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2
asyncpg
pydantic-settings
greenlet
python-jose[cryptography]
requests
python-multipart
bcrypt>=4.0.0
websocket-client

# 模块六：音频处理
resemblyzer
numpy
librosa
soundfile
